throughout the wing and the stage i alone am the autonomous one
