--- src/main/java/frc/AutoRoutine/StartOnLeft.java
@@ -1,7 +1,0 @@
-package frc.AutoRoutine;
-
-import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-
-public class StartOnLeft extends SequentialCommandGroup {
-    
-}
